+++ Dyalect.Console/DyaOptions.cs
@@ -1,5 +1,8 @@
 ﻿using Dyalect.Util;
+using System;
+using System.Linq;
 using System.Collections.Generic;
+using System.Text;
 
 namespace Dyalect
 {
@@ -21,13 +24,49 @@ namespace Dyalect
         [Binding("path", Help = "A path where linker would look for referenced modules. You can specify this switch multiple times.", Category = LINKER)]
         public string[] Paths { get; set; }
 
-        [Binding("nocol", Help = "Disable colors in console.", Category = GENERAL)]
-        public bool NoColors { get; set; }
-
         [Binding("nologo", Help = "Do not show logo.", Category = GENERAL)]
         public bool NoLogo { get; set; }
 
-        [Binding("theme", Help = "Sets a name of Dyalect Console visual theme.", Category = GENERAL)]
-        public string Theme { get; set; }
+        [Binding("i", Help = "Stay in interactive mode after executing a file.", Category = GENERAL)]
+        public bool StayInInteractive { get; set; }
+
+        public override string ToString()
+        {
+            var list = new List<(string, string)>();
+            
+            foreach (var pi in typeof(DyaOptions).GetProperties())
+            {
+                var attr = Attribute.GetCustomAttribute(pi, typeof(BindingAttribute)) as BindingAttribute;
+
+                if (attr != null)
+                {
+                    var val = pi.GetValue(this);
+                    var byt = pi.PropertyType == typeof(bool);
+                    var i4 = pi.PropertyType == typeof(int);
+
+                    if ((byt && !(bool)val)
+                        || (i4 && (int)val == 0)
+                        || ReferenceEquals(val, null))
+                        continue;
+
+                    var key = attr.Names?.Length > 0 ? attr.Names[0] : "<file name>";
+                    list.Add((key, byt ? "" : val.ToString()));
+                }
+            }
+
+            var max = list.Max(e => e.Item1.Length) + 1;
+            var sb = new StringBuilder();
+
+            foreach (var (k,v) in list)
+            {
+                if (sb.Length > 0)
+                    sb.AppendLine();
+
+                sb.Append(("-" + k ).PadRight(max, ' ') + " ");
+                sb.Append(v);
+            }
+
+            return sb.ToString();
+        }
     }
 }