+++ Dyalect.Console/Util/CommandResult.cs
@@ -1,9 +1,0 @@
-﻿namespace Dyalect.Util
-{
-    public enum CommandResult
-    {
-        None,
-        Exit,
-        Reset
-    }
-}